[FIRRTL] Fix lowering of the firrtl.connect to rtl.connect op.
The firrtl.connect op does implicit extension, so we need to handle
this before generating an rtl.connect.

Unfortunately, this exposes the fact that we cannot use the standard MLIR
lowering pass for this transformation (well, unless we're willing to introduce
tons of ugly casts and eliminate them later).  I added a FIXME about this, but
will rewrite the pass in a seperate step.