[FIRRTL] Add Circuit Op Verification (#88)
* [FIRRTL] Fix tests so "main" matches a module

Fix FIRRTL tests to make them valid FIRRTL IR. Previously, the example
FIRRTL IR or FIRRTL MLIR dialect examples could contain a circuit
whose name (what FIRRTL IR calls the "main" or, equivalently, the top
module) did not match the name of a module in the circuit. This commit
changes the tests so that this invariant holds.

* [FIRRTL] Add circuit verifier

Add a new function, verifyCircuitOp, that is used to verify a FIRRTL
circuit op. This checks the following two conditions:

  1. The circuit must have a name ("main" in FIRRTL IR/the top module)
  2. The circuit must contain a module matching the circuit name

This brings the FIRRTL dialect in line with the current implementation
of the FIRRTL compiler where condition (1) is nonsensical and
condition (2) is rejected.

* [FIRRTL] Test circuit verifier

Add a test for the FIRRTL dialect that the user is provided with an
error if they provide FIRRTL MLIR whose name (it's "main" or top
module) does not match the name of any modules in the circuit.

Add a test that the FIRRTL dialect circuit op has a non-empty name.