[FIRRTL] Improve RegOp/RegInitOp Verification (#95)
* [FIRRTL] Correct RegInitOp description

Fix copy-paste typo where the description of RegInitOp uses the "reg"
keyword instead of "reginit" keyword.

* [FIRRTL] Enforce register op clock is a clock type

Change the type of the clock connection for RegOp and RegInitOp from
AnyType to ClockType.

* [FIRRTL] Add checks the register clock is a clock

Add tests for RegOp and RegInitOp that the type of the clock
connection is actually a clock.

* [FIRRTL] Add FIRRTLType.isResetType

Add a method that determines if a FIRRTLType is a valid reset type. A
valid reset is one that is either an abstract reset, a concrete 1-bit
UInt, or a concrete asynchronous reset.

* [FIRRTL] Add constraints to ResetType

Use new FIRRTLType::isResetType to assert that a ResetType is actually
a valid reset type.

* [FIRRTL] Require RegInitOp reset is ResetType

Change the required type of the reset connection in a RegInitOp from
AnyType to ResetType. This relies on constraint checking of ResetType
and will accept any abstract reset, concrete asynchronous reset, or
concrete 1-bit UInt.

* [FIRRTL] Test rejection of bad RegInitOp resets

Add a test that a 2-bit UInt reset connection is rejected as invalid
FIRRTL.