[RTL] Improve repr and EmitVerilog for parameterized module instances.
This removes the unused parameter dict from rtl.externmodule -
parameters in Verilog don't live on the declarations of modules,
they live on instances.

This adds a parameter dict to the rtl.instance op, adds EmitVerilog
support for them, and teaches the FIRRTL LowerToRTLModule pass to
handle this (more) correctly.

There is still a lot of work to be done, I'll file some follow-up
bugs.