[ESI] [Cosim DPI] Integration test framework (#228)
* [ESI] [Cosim DPI] Integration test framework

Number 5 of 7 planned PRs for #213. The focus of this PR is a lit format
for Cosim. Cosim integration tests require selecting a TCP port,
starting the simulator (which opens a server on that port), then running
the test(s). This lit format enables tests written in Python.

This PR might be difficult to understand without the actual test -- which
fails since the Cosim DPI plugin has not yet made it into
master. I can add the test after we've PR'd the rest of the cosim DPI
code.

* An unrelated change made it in.

* Consistent casing

* Check for pycapnp also