[FIRRTL] Handle flipped bundle types in LowerTypes (#263)
The canonical representation of a BundleType with every field flipped
is a FlipType that wraps the whole bundle. There were places where
LowerTypes assumed a BundleType, which caused crashes for
a FlipType wrapped around a BundleType. To address this:

* LowerTypes will default to disabled in firtool, and can be enabled by
passing the -enable-lower-types flag.
* The instance op verifier was updated to guarantee the result type is a
BundleType or a FlipType wrapped around a BundleType. 
* LowerTypes was updated to check for both scenarios in the places it
previously assumed it was working with a BundleType.