[RTL] Change rtl.connect to require an inout type for its destination.
This allows us to more directly model the invariant that you can only connect
to an inout entity like a port or wire.

The majority of this change is updating testcases, as well as a significant
update to the LowerToRTL pass.  That pass is updated to preserve inout types
for analog and inout ports.  The LowerToRTLModule pass is also changed to
generate firrtl.wire instead of rtl.wire nodes when handling ports that are
non-obvious.  This made the logic simpler.