[LowerToRTL] Fix issue #373.
Avoiding wire insertion for instance inputs is non-trivial: the connects
to the input happen lexically after the firrtl.instance, but the operands
to an rtl.instance need to be lexically before the rtl.instance.  There
are also some pretty crazy cases in which such a transformation will
introduce cycles.

While the RTL dialect supports cyclic circuits, the LowerToRTL pass does
not.  To solve these problems, scoot the computation that feeds an
input lexically ahead of the rtl.instance when possible, and just bail
out and insert a wire when not possible.