[FIRRTL] Canonicalize away flip of analog types.
This implements the approach described in Issue #324, whereby we turn
flip(analog) into analog.  Analog types in FIRRTL are implicitly
bidirectional already, so an "output port of analog type" doesn't
actually make sense.  This canonicalization defines away certain
defensive code and eliminates the need for "AsPassive" to apply to
values of analog type.

The FIRRTL version of the verilog printer now prints analog ports as
inputs, if someone cared we could make it properly print inout for
them.