[FIRRTL] Verify that NodeOp is passive (#416)
During parsing, add an asPassive cast to a node argument if it is
passive under an outer flip. Error if type is analog or underlying
type is not passive. Restrict a NodeOp's input to be passive.

Enables parsing of weird CHIRRTL circuits where a node is a "sink" (an
output port or an instance input).

Add test cases of this behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>