[LLHD] Add connect operation. (#554)
This operation is unfortunately necessary to model structural
connections between signals. The assembly format is modified from the
LLHD spec to print the type on the right like other LLHD MLIR
operations. The operation name in C++ is ConnectOp to align with other
dialects' choice of name for this concept.