[ExportVerilog] Make "out of order expr handling" aware of transitive uses.
Issue 595 pointed out a testcase where multiple inlinable expressions can
happen in a use-chain that causes something to get emitted out of line.
Because of this, we should emit a declaration of a wire and then a later
assign into it, intead of emitting the wire declaration inline.  Fix this
by tracking all the expressions that are transitively out of order in a
prepass (keeping linear time) and referencing it later.

Before we'd emit:

  assert(_T[6'h0+:32] == 32'h0);
  wire [63:0] _T = arr[7'h0+:64];

now we emit:

  wire [63:0] _T;
  assert(_T[6'h0+:32] == 32'h0);
  assign _T = arr[7'h0+:64];