[FIRRTL] Remove use of FunctionLike helpers in LowerTypes. (#651)
The previous implementation relied on the setArgAttrs helper to copy
over the name attribute for every new argument added to the
block. Instead, this accumulates an array of attributes, and sets all
of them in one shot at the end.

Besides setArgAttrs, this also removes the eraseArguments helper. The
performance is subpar compared to directly setting all the attributes
at the end. This means we need to manually remove the block arguments,
and directly set the module's type attribute, instead of relying on
the helper.

With these changes, we take over the bookkeeping work done by the
upstream MLIR helpers to improve performance. The upstream helpers
have to be very careful to keep attributes in a consistent state;
by doing it ourselves, we improve performance by leaving the attributes in an
inconsistent state until the end of the pass.