[FIRRTL LowerTypes] Add FExtModule type lowering (#743)
[FIRRTL LowerTypes] Add FExtModule type lowering

This adds FExtModule lowering support to the lower types pass. This
flattens all types in the parameter list and gives each new operand a
new name.

To process ExtModules in this pass, it had to change from a FModuleOp
pass to a CircuitOp pass. In order to maintain performance, this pass is
now parallelized internally using `parallelFor`.