[ExportVerilog] Replace `.` in names with `_`
Previously a `.` was replaced with its hex value, which is `2E`. This
change makes it so `.` are replaced with `_` characters. Since we are
purposefully generating names with `.` in them, it makes sense to
replace them with something more readable than the hex value.

This change also moves the check for a valid identifier starting
character until after replacing all illegal verilog characters.
Previously, an identifier starting with a `.` would be prefixed with an
`_` and then the `.` would be replaced with a `_`.  This produced an
identifier with two leading `_` characters.