[FIRRTL] Restructure single-connect folding to be around the connect.
This is mostly a merging and unification of the reg/wire folding logic,
but moves to being centered on the connect. This will work better when
we want to handle mixed type connects (by making the extension explicit).

This also adds support for port forwarding through wires.  This is part
of Issue #794 but is not itself enough to close it off.