[ESI] Lower ESI Channel ports on modules defined in CIRCT to data/valid/ready (#835)
For external modules, ESI lowers to SV interfaces since it assumes a user is going to write a module against that generated SV interface. For internally-defined modules, lowering to regular data/valid/ready wire signaling is preferred so as to not preclude further optimization. SV interfaces also bloat the IR.