[ExportVerilog] Change line wrapping length policies to be consistent.
We have a "preferredSourceWidth" setting that controls the target width
of the generated verilog that port printing was using, and different
magic numbers that expression splitting was using.  Pull them into a
common regime.

This isn't making a statement that "90" is the right answer, just that
one knob is better than 3. :-)