[ExportVerilog] Pull the "out of order use" handling logic into a prepass.
Trying to handle it inline in the verilog emitter was too complicated and
not working well.  Pull it into a prepass (which generates explicit wire
declarations to resolve forward refs/cycles as needed) to keep the logic
simple.

This fixes a bug where wires would get emitted multiple times and Issue #838.