[RTL] Store port names more efficiently in rtl.module and friends.
This also improves the printing and parsing of ext modules and other
things to print with pretty syntax, and eliminates the rtl.name
attribute.

This new design is MUCH faster, uses less memory, and is simpler to
work with for passes.  It eliminates a bunch of dictionary attr with a
single rtl.name key in it, and gets us out of making a bunch of arg1
attributes on the module itself.

This is the RTL half of Issue #872.