[FIRRTL] Add default name Reg/RegResetOp Builders
Add builders to RegOp and RegResetOp that include a default String
name argument.  This cleans up the API for creating registers where a
user doesn't have to create an attribute and can use a String.  Also,
this will make it easier to make modifications to these ops in the
future, e.g., when adding annotation support.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>