[Seq] Computational register op and lowering to SV (#883)
Model of the compreg op outlined in the Seq dialect rationale dialect. This is just a starting place and I do expect it to change and/or have siblings for different flavors. Lowers to sv.alwaysff with posedge clock and sync reset -- for now.

Also adds a few mechanical things required to make the Seq dialect actually work which I neglected in the PR which added the dialect.