[Handshake] Fix a bug in control merge canonicalization.
This previously violated the rule that "no IR mutation should take
place before the match is deemed successful".

This comes up in larger graph canonicalizations, and would lead to the
error "The pattern rewrite doesn't converge after scanning 10
times". The new test case isn't enough to catch this complex
situation, but it does cover this change.

https://mlir.llvm.org/docs/PatternRewriter/#match-and-rewrite-implementation