[FIRRTL] Eliminate the portNames attribute on firrtl.instance.
This is a redundant encoding of information that already exists on the
referenced module, so just use the referenced module.  This simplifies
a bunch of code and makes things faster, because we now do association
between instance and module ports by index instead of by name.  This
should also save a bit of memory.