Remove Type Canonicalization (#944)
Remove FIRRTL Type Canonicalization.  Strip flips from FIRRTL
operations, e.g., subfield "a" of type "bundle<a: flip<uint<1>>>" is
now "uint<1>".

This works towards moving the FIRRTL Dialect representation of FIRRTL
IR to be as close as possible to the Scala FIRRTL Compiler
implementation.

* [FIRRTL] Add UnflippedType, use for WireOp result

Add an UnflippedType to the FIRRTL Dialect.  This is a type which does
not contain an outer flip (but may contain internal flips).  This
accurately represents the legal types of wires.

Change WireOp to use UnflippedType for its result.

* [FIRRTL] Add SubfieldOp::isFieldFlipped Utility

Add a utility that returns true if the field of a subfield is
flipped.  Because the type of a subfield now strips the flip (to align
with the Scala FIRRTL Compiler implementation), this utility is useful
when you need to know the flip, e.g., when computing flow.

* [FIRRTL] Add basic utilities for Flow

Adds basic utilities for working with FIRRTL flow.  A firrtl::Flow
enum is added as well as a foldFlow function that will compute the
flow of a Value.

* [FIRRTL] Add basic kind utilities

Add enums for representing necessary FIRRTL kinds.  Add a utility,
getDeclarationKind that returns the kind of a value.

* [FIRRTL] Use Flow/Kind ConnectOp Verification

* [FIRRTL] Support LowerTypes on Canonicalized Types

* [FIRRTL] LowerToRTL for Un-Canonicalized Types

* [FIRRTL] Parse 'is invalid' for new types

Change parsing for `is invalid` to emit expanded, leaf connects of
sink or duplex destinations.

* [Handshake] HandshakeToFIRRTL for un-canonicalized types

- FModule lowering has been fixed to use an outer flip type for output
  parameters.

- Subfield op has switched to using a builder which calculates the
  proper result type automatically.  This is the source of most test
  changes.

- Some type handling has been switched to handle an outer flip around
  bundle types.

Co-authored-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Co-authored-by: Andrew Young <youngar17@gmail.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>