[FIRRTL][LowerTypes] Don't read port names when lowering InstanceOps
This fixes a data race in LowerTypes.  While lowering InstanceOps, the
port names were read from the ModuleOp and used while flattenning bundle
typed results, while another thread modified the ModuleOp's port names.

This was a race condition which lead to to instance results not
correctly being replaced with flattened results, or dropping too many
characters from the fron of a bundle field name.

The port name was not required to flatten the bundle, and was later
stripped from the flattened bundle. This change stops using the port
name while flattening the instance results, and an empty string was used
instead.