[RTL][Python] Add InstanceBuilder for constructing instances. (#940)
Python. The builder accepts a name for the instance and optional
values to assign to the input ports. The InstanceBuilder supports
attribute getters and setters for the ports.

When ports are not set on creation, temporary operations are inserted
into the IR via a BackedgeBuilder. The BackedgeBuilder can be used as
a context manager, and checks the temporary operations are all
resolved on exit. This is wrapped around the body_builder of modules
to ensure all instances are wired up after the body_builder exits.

A create method is added to RTLModuleOp, which constructs and returns
an InstanceBuilder for an instance of that module.