Support ReferenceTarget Annotations for Ports (#882)
Add support to the FIRRTL Dialect for parsing, printing, and storing
ReferenceTarget Annotations that target ports.  These annotations are
stored in arg attributes as they are expected to be sparse.

Update LowerTypes to split these annotations when ports are split.
Other arg attributes are preserved (copied to all lowered ports).  (No
other arg attributes exist, but this will add some future proofing to
LowerTypes.)

Add tests of this behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>