Support Scattering SiFive "Grand Central" Annotations (#1056)
Add in-tree support for scattering known, custom FIRRTL Annotations to
disparate targets.  A global annotationID is added to the parser's state
that links scattered annotations.

Add support for scattering:
  - sifive.enterprise.grandcentral.DataTapAnnotation
  - sifive.enterprise.grandcentral.MemTapAnnotation
  - sifive.enterprise.grandcentral.GrandCentralView$SerializedViewAnnotation

During scattering, verify the correctness of the Annotations that are
being processed.  Annotations which fail verification will result in
verbose errors about what field was missing or that a type was
incorrect.  Previously, this would fail with casting errors or
llvm_unreachable.

Miscellaneous additions:
  - Add support for parsing AugmentedType objects
  - Add tests of the three custom annotations listed aboe
  - Add one test of a verification error

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>