[FIRRTL] Fix Regression on Mux w/ Known Width
Fix a regression introduced in 2b84ba55 for width inference involving
the Mux select line.  2b84ba55 was yolo-ing a cast causing width
inference to crash if it ran on a Mux with a select line that had a
known width.  tl;dr: I should learn how to use the APIs...

Fix #1108 (again).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>