[FIRRTL][LowerTypes] Report the op when missing a bundle lowering (#1111)
This will print an error message when a bundle lowering does not exist.
This is usually the result of missing support for a specific operation
in the LowerTypes pass.  It is helpful to know what operation support is
missing without having to start up the debugger.  In order for the error
message to reliably print, multithreading must be disabled with:
`--mlir-disable-threading`.

Example output:
```
./fieldtest.fir:6:5: error: bundle lowering was not set
    node n = in0
    ^
./fieldtest.fir:6:5: note: see current operation: %n = firrtl.node %arg0  : !firrtl.bundle<a: uint<1>>
LLVM ERROR: bundle lowering was not set
PLEASE submit a bug report to https://bugs.llvm.org/ and include the crash backtrace.
Stack dump:
0.	Program arguments: ./bin/firtool -mlir-disable-threading ./fieldtest.fir
```