[Python] Set a default context (and location) upon import (#1107)
The most common use case for the `circt.design_entry` API is that a single
context is used throughout the lifetime of the application. This creates a
default context and default location and installs them into the defaults stack.

Since we don't want this behavior whenever someone imports circt, don't import
`design_entry` into the `circt` namespace.