[Python] Add generic connect API that works with builders. (#1106)
This adds `circt.connect` to "connect" two values. It sets the
destination OpOperand to the new value.

To integrate with the builder approach, this updates the
InstanceBuilder to return a wrapper BuilderValue around any Values
accessed through `__getattr__`. This works with the connect API to
pass around the needed index to ultimately set the Value and erase any
temporary backedges the builder created.

As the builder approach is generalized to other ops, the connect API
should work in exactly the same manner.