[Py] Make @module annotations create MLIR Operations (#1123)
Make the @module annotation create a subclass which also inherits from OpView. When OpView is initialized, an unregistered MLIR operation gets created. Use the Input and Output classes as classmembers or instancemembers to create operands and results.

This breaks the construct based building. That'll be added back once we have a lowering API in python.