[Python] Add NamedValueOpView builders for the Comb dialect. (#1163)
This refactors the NamedValueOpView to push all the complexity into a
shared constructor that works for a variety of ops like InstanceOp,
CompRegOp, and the non-variadic Comb ops.

To support this, subclasses must now define two hooks to retrieve the
list of operand names and result names. There is also a third hook to
create a backedge of an appropriate type for a given operand that was
left uninitialized. The default implementation works for ops that have
the same type for all operands.