[PyCDE] Add generator options to select a generator by name. (#1229)
The `generate` call accepts a list of candidate generators to run,
which defaults to empty.

When looking up a generator, the list is consulted. If it is empty,
and there is one generator, that is used by default. If it is present,
and it names one of the candidate generators, that is selected. If it
names multiple candidate generators, an error is raised.