[FIRRTL][BlackBoxReader] Stop emitting duplicate blackboxes  (#1225)
Before this change, BlackBoxReader will append multiple BlackBox
annotations with the same filename to the same output file. This is
different than the SFC behaviour, which will emit only one of the
existing blackbox annotations to the output file.

This is causing problems when there is the a duplicate BlackBox
annotation in the IR, which is resulting in multiple definitions of
modules in the output. Duplicate blackbox resource annotations are
fairly common, for example plusarg_reader is included from many
places.

The current firtool behaviour is a result of how these annotations are
lowered: each BlackBoxAnnotation becomes an sv.verbatim node, which
each append their contents to the destination file. This change updates
the BlackBox reader to only keep and lower the first time it sees a
black box resource.

There is no checking that BlackBoxAnnotations targeting the same file
are identical in contents. This is the SFC behaviour, although we might
consider revisiting this.