[FIRParser] Fix crash when mport is eating the skip op (#1249)
With the new lazy-location tracking, it is important that certain
location tracking book keeping be handled before and after every
statement.  The `mport` op parsing has a hack to parse the next
statement when it is a `skip` operation, but was not doing to correct
set up to parse it.