[TranslateToVerilog] Fix some twine lifetime issues, NFC.
Twines are a very sharp tool for avoiding string allocations.  Their
lifetime depends on temporary objects being alive when the twine is
evaluated.  This means that they are really only safe to use as a
subexpression in an argument, they can't be used safely as return
values or when split across multiple C++ statements.