[FIRRTL] Cleanup asPassive casting in parser (#1262)
Change all asPassive casting in the FIRRTL parser to introduce a cast if a value
is not passive or if it has sink flow.  This unifies some of the whack-a-mole
that was introduced in 3a021616 and continued in fcc69b51.

Simplify the parser, by changing convertToPassive to also do flow checking as
defined above.  This then allows forcePassive (introduced in 3a021616) to be
removed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>