[FIRRTL] Simplify and speed up annotation support in parser.
It is simpler to just return the AnnotationSet instead of filling it in,
the one place that needs to merge them can do so explicitly.  This also
changes the getAnnotations() method to use a SmallString internally instead
of an std::string to avoid an allocation on every lookup (when an annotation
set is available).

This also makes MemoryPortOp annotation printing consistent with the rest of
the ops, but there are other problems here.