[FIRRTL] Eliminate FlipType (#1309)
FlipType doesn't propagate the same way Flow does, and becomes a redundant way to try to represent FIRRTL's flow concept. This matches FIRRTL by attaching flip as a property of bundle fields. This eliminates a fair amount of code which constantly was looking around flips. Only two operations need special consideration: InstanceOp and MemOp. When computing flow for an InstanceOp result, you need to look at whether it was an input or output for the referenced module. For MemOp, all flows are sinks.