[FIRRTL] Add Support for Instance Annotations
Add basic support for Annotations that have InstanceTargets.  These are
Annotations that target exactly one level of hierarchy.  This style of
annotation is fungible with a ReferenceTarget pointing at an instance
and should be treated the same in CIRCT.

E.g., the following targets instance "bar" of module "Bar" in module
"Foo".  It then targets something in "Bar" named "a".

    ~Foo|Foo/bar:Bar>a

This can be equivalently expressed as the following ReferenceTarget:

    ~Foo|Foo>bar.a

Add support for parsing these and adding them as attributes on either
instances or their ports.  The restriction on parsing is now changed to
truly disallow non-local annotations.  I.e., any annotation with a
target that has _more than_ one level of hierarchy will cause a parsing
error.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>