[FIRRTL][Annotation] Support port annotations for MemOp (#1317)
As `MemOp` can have multiple ports, such as `r` and `w`, we want to add
annotations only applicable to one of the ports. Similar to `InstanceOp`, this
patch stores the port annotations of `MemOp` in an `ArrayAttr` named
`portAnnotations` and supports to access them through the port index. This patch
also updates the `LowerTypes` pass by filtering the port annotations and
attaching applicable annotations to the lowered `MemOp`s.