[HW] Support type aliases where ArrayType is expected. (#1143)
This is accomplished with a new set of isa/cast/dyn_cast helpers in
the circt::hw namespace. These helpers build on the normal MLIR
helpers of the same names, but additionally know how to see through a
type alias to its canonical type.

The hw::ArrayType in C++ is unchanged, but in ODS, ArrayType is
defined in terms of the new helpers. This allows all ODS ops that work
with an ArrayType to now work with an ArrayType or a type alias
wrapping an array type.

The various predicates and parsers are updated to use the new helpers.