[FIRRTL] Simplify subTarget utility for Annotations
Change an Annotation utility that is used for splitting out subindex and
subfield tokens from a Target string to also update a prototype
Annotation if any subfield or subindex tokens are found.  This saves
duplicate code in multiple places and will help in the future as this
new utility (and annotation update pattern) is going to be used
everywhere during Annotation scattering.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>