[FIRRTL] Fix DontTouch for ReferenceDataTapKey
Fix a bug in generation of DontTouchAnnotations for ReferenceDataTapKeys
that included subfield/subindex tokens.  Previously, the entire,
unmodified Target would be passed to the FIRRTL Parser (and it would
drop the Target thinking it didn't apply to anything in the circuit).
E.g., the target would be stored as "~Foo|Foo>bar.baz.qux".  For this to
work with the parser, this should be keyed with "~Foo|Foo>bar" and the
Annotation body should include "target = ['.baz', '.qux']".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>