[FIRRTL][LowerToHW] Handle 0-bit mux sel (#1349)
Add a new type, UInt01Type, which is a UInt that has either a 0-bit,
1-bit, or unspecified width.  Change the mux primop to use this type for
its select line (since this is what the Scala FIRRTL Compiler allows).

Add a folder that will fold a mux with a 0-bit select line to its false
value.  This aligns CIRCT with the behavior of the ZeroWidth pass in the
Scala FIRRTL Compiler (which will replace a 0-bit select with a 1-bit,
0-valued select) along with its constant propagation behavior.

Handle FIRRTL MuxPrimOp's with 0-bit select lines during FIRRTL to HW
lowering.  Previously, these were unhandled and would produce errors.
Lower these by choosing the false path and padding if necessary.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>