[Calyx] Add guard for assignments and done operations. (#1366)
Adds guards to Calyx assignments and done operations. A guard is
a boolean expression that determines when the given assignment
should be active. The format of the guard is slightly different than
that of the original Calyx IR, to avoid parsing an optional guard as 
a source.
Original IR: `dest = guard ? src`
Dialect: `dest = src, guard ?`