[Calyx] Add Enable and Sequential Ops. (#1372)
Adds EnableOp and SequentialOp to the Calyx dialect.

This is a first attempt at Calyx's control section. I've used a symbolic representation (similar to the current Calyx IR), so that groups can be verified and accessed by name. The control regions inherit from CalyxContainer given their
simple specification, but may diverge in the future.