[Calyx] Clean up ControlLike trait. (#1395)
This PR cleans up ControlLike. The trait should only focus on operations found within the
ControlOp's body. The verifier returns successfully early if the operation does not have a body. 
This should make it easier to add future control operations with and without bodies, such as 
(the future incoming) InvokeOp and ParOp. Also adds a verifier is added to ControlOp to allow for 
single EnableOps within the body.