[FIRRTL][InferWidths] Fix unconstrained widths inferred to zero
Fix an issue in `InferWidths` where an unconstrained width could be
inferred to zero. Curiously, this only happens if the unconstrained
width is used as part of another constraint expression, e.g. a connect
of one unconstrained width to another. This fix makes the width
calculation properly return a `None` value instead of `0` for the width.
Also fixes a nasty subtle bug where the cycle breaking code was not
properly balanced.

Fixes #1381.