[ExportVerilog] Emit bind expressions.   (#1414)
Emit bind expressions.  This uses the fact that prior to emission, prepare has ensured that all arguments to an instance used by bind are tied to nameable things (wires, regs, ports) and those names are legalized.  This allows emission to pull names from wires and ports and use them from the top level or another module.