[FIRRTL] Use SmallVector for Module Storage, NFC
Change from std::vector<A> to SmallVector<A, 0> for storing deferred
modules that need to be parsed.  There are some issues that seem to be
related to the exception-safe nature of std::vector that need a copy
constructor when it's getting implicitly deleted for certain versions of
glibc.  This started to show up with using allocators in members of the
DeferredModuleToParse struct.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>