[ExportVerilog] Add option to force wires in event control exprs
Add the `exprInEventControl` lowering option to `ExportVerilog`, which
if forces the emitted Verilog to spill constants (which live in a
`localparam`) into a `wire` when they are used as part of an event
control (`@(posedge ...)`). The `exprInEventControl` option restores the
old behavior of allowing arbitrary expressions in the event control.

Fixes #1450.