[FIRRTL] Fix up firtool integration tests
This updates the integration test to make every module reachable from
the top-level module.  This prevents unused modules from being deleted
by one of the optimizations.

Some of the `CHECK :` lines were non-functional because of some extra
white space. These lines were fixed and the tests were updated to match
the current output.

There was some shuffling of the test output to try to move the `CHECK`
lines closer to the FIRRTL module definition, and some check lines  were
rearranged to be in pipeline order `FIRRTL -> HW -> Verilog`.