[SV][ExportVerilog] Add Ability to Bind Interfaces (#1432)
* [SV] Add BindInterfaceOp

Add a new op, "sv.bind.interface", that can be used to bind an
interface.  This op is symmetric with respect to BindOp ("sv.bind") and
provides the same functionality, but operates on SV interfaces instead
of HW instances.

* [SV] Add name, symbol to InterfaceInstanceOp

Add a mandatory name and an optional symbol to InterfaceInstanceOp.
Derive the name from the SSA value if one is available.

* [ExportVerilog] Add naive BindInterfaceOp emission

Add basic emission of BindOpInterface.  This emits using the janky
"(.*)" approach (where anything matching is connected).  This is
non-ideal, but interfaces are not as straightforward to handle as
instances where all connections are explicit.

This is intended to be a temporary emission strategy that can be
improved in the future.

* [ExportVerilog] Split interface instance emission

Move interface instance emission out of the collectNames and
getVerilogDeclWord emission path and get this into its own method (like
InstanceOp).

Add conditional emission gated on the presence of a "doNotPrint"
attribute, just like InstanceOp.

* [ExportVerilog] Add test of bind interface, NFC

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>